feat: add otpApi service
- renamed sendOTPApi to otpApi
- added verifyOTP service to otpApi
